backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-chroma
sentence-transformers      
requests
python-dotenv
